manually bootstrap eval
instead of using synthetic data to ensure quality

+++ eval/tests/eval_test.py
@@ -1,7 +1,7 @@
 import pytest
 from pytest import fail
 import subprocess
-from eval.types import TroubleshootingQuestion
+from eval.types import TroubleshootingQuestion, QNA, VerificationStep
 import yaml
 import structlog
 import tempfile
@@ -14,13 +14,14 @@ from opsmate.libs.core.types import Agent, ReactAnswer
 from openai import OpenAI
 import instructor
 from pydantic import BaseModel, Field
+import time
 
 logger = structlog.get_logger()
 
 
 def issues() -> list[TroubleshootingQuestion]:
-    with open("./eval/issues.yaml", "r") as f:
-        return [TroubleshootingQuestion(**issue) for issue in yaml.safe_load(f)]
+    with open("./eval/q_n_a.yaml", "r") as f:
+        return [QNA(**issue) for issue in yaml.safe_load(f)]
 
 
 @pytest.fixture
@@ -39,10 +40,15 @@ def using_eval_cluster():
 
 
 @pytest.fixture
-def with_namespace(issue: TroubleshootingQuestion):
-    subprocess.run(["kubectl", "create", "namespace", issue.namespace], check=True)
+def with_env(issue: QNA):
+    if issue.namespace is not None:
+        subprocess.run(["kubectl", "create", "namespace", issue.namespace], check=True)
     yield
-    subprocess.run(["kubectl", "delete", "namespace", issue.namespace], check=True)
+    if issue.namespace is not None:
+        subprocess.run(["kubectl", "delete", "namespace", issue.namespace], check=True)
+
+    for step in issue.cleanup_steps:
+        subprocess.run(step.command.split(), check=True)
 
 
 @pytest.fixture
@@ -62,7 +68,7 @@ def executor():
     return AgentExecutor(client=OpenAI())
 
 
-class RootCauseScore(BaseModel):
+class OutputScore(BaseModel):
     score: int = Field(
         description="The score between 0 and 100 based on how similar the actual output is to the expected output",
         ge=0,
@@ -71,8 +77,8 @@ class RootCauseScore(BaseModel):
 
 
 def verify_root_cause(
-    question: str, candidate_answer: str, expected_root_cause: str
-) -> RootCauseScore:
+    question: str, candidate_answer: str, expected_output: str
+) -> OutputScore:
     cli = instructor.from_openai(OpenAI())
     return cli.chat.completions.create(
         model="gpt-4o",
@@ -86,9 +92,9 @@ You are a sysadmin examiner tasked to verify whether the actual root comes up fr
 {question}
 </Question>
 
-<Expected Root Cause>
-{expected_root_cause}
-</Expected Root Cause>
+<Expected Output>
+{expected_output}
+</Expected Output>
 
 <Candidate Answer>
 {candidate_answer}
@@ -98,28 +104,29 @@ Please give a score between 0 and 100 based on how similar the candidate's answe
 """,
             }
         ],
-        response_model=RootCauseScore,
+        response_model=OutputScore,
     )
 
 
 @pytest.mark.parametrize("issue", issues())
 def test_load_issues(
-    issue: TroubleshootingQuestion,
+    issue: QNA,
     using_eval_cluster,
-    with_namespace,
+    with_env,
     supervisor: Agent,
     executor: AgentExecutor,
 ):
-    for step in issue.steps_to_create_issue:
-        # write the manifest to a temp file
-        with tempfile.NamedTemporaryFile(delete=False) as f:
-            f.write(step.manifest.encode("utf-8"))
-            f.flush()
-            subprocess.run(["kubectl", "apply", "-f", f.name], check=True)
+    # for step in issue.steps_to_create_issue:
+    #     # write the manifest to a temp file
+    #     with tempfile.NamedTemporaryFile(delete=False) as f:
+    #         f.write(step.manifest.encode("utf-8"))
+    #         f.flush()
+    #         subprocess.run(["kubectl", "apply", "-f", f.name], check=True)
 
     supervisor_output = executor.supervise(
         supervisor,
-        f"In the {issue.namespace} namespace, {issue.question}",
+        # f"In the {issue.namespace} namespace, {issue.question}",
+        issue.question,
     )
 
     for output in supervisor_output:
@@ -128,9 +135,58 @@ def test_load_issues(
             break
 
     # makes sure the output is similar to the root cause
-    score = verify_root_cause(
-        question=issue.question,
-        candidate_answer=output,
-        expected_root_cause=issue.root_cause,
+    if issue.answer_command:
+        # execute the command and verify the output
+        expected_output = subprocess.run(
+            issue.answer_command.split(),
+            stdout=subprocess.PIPE,
+            stderr=subprocess.STDOUT,
+            text=True,
+        )
+        score = verify_root_cause(
+            question=issue.question,
+            candidate_answer=output,
+            expected_output=expected_output,
+        )
+        assert score.score > issue.similarity_threshold * 100
+
+    for verification in issue.answer_verification:
+        # execute the command and verify the output
+        # result = subprocess.run(
+        #     verification.command.split(),
+        #     stdout=subprocess.PIPE,
+        #     stderr=subprocess.STDOUT,
+        #     text=True,
+        # )
+        # assert result.returncode == verification.exit_code
+        # assert result.stdout == verification.expected_output
+        for _ in wait_until():
+            verify_output(verification)
+
+
+def verify_output(verification: VerificationStep):
+
+    result = subprocess.run(
+        verification.command.split(),
+        stdout=subprocess.PIPE,
+        stderr=subprocess.STDOUT,
+        text=True,
     )
-    assert score.score > 60
+    output = result.stdout
+    exit_code = result.returncode
+    assert exit_code == verification.exit_code
+    assert verification.expected_output in output
+
+    return True
+
+
+def wait_until(timeout: int = 10, period: int = 1):
+    mustend = time.time() + timeout
+    while time.time() < mustend:
+        try:
+            yield
+            return
+        except AssertionError as e:
+            pass
+        time.sleep(period)
+    raise TimeoutError(f"Timeout after {timeout} seconds")
